feat: Added support for snowflake and bigquery databases. Optimized the deployment pattern for lambda,in order to slove the problem with lambda quotas
