--- services/connector/lambda/env.json
@@ -1,5 +1,0 @@
-{
-  "connection_string": "clickhousedb://<user>:<password>@<host>:<port>/<database>[?options…]clickhouse://{username}:{password}@{hostname}:{port}/{database}",
-  "database": "databases",
-  "table": "table"
-}
